Rewrite test so its more resilient to DNS issues.
Have occasional failures when run on CPAN Testers due to DNS lookup
issues, which caused the "-mxcheck" to fail when trying to constrain
results to valid e-mail addresses.

Rewrite the test suite so we can do an initial check to see if we have a
good DNS setup to test against, and to skip those constraint tests if we
do not.